refactor: 멤버 관련 테스트 클래스에 대한 패키지 구조 변경
