+++ src/main/java/kr/binarybard/hireo/web/member/domain/Member.java
@@ -1,4 +1,4 @@
-package kr.binarybard.hireo.member.domain;
+package kr.binarybard.hireo.web.member.domain;
 
 import lombok.Builder;
 import org.springframework.security.crypto.password.PasswordEncoder;
