+++ src/main/java/kr/binarybard/hireo/web/member/domain/Role.java
@@ -1,3 +1,3 @@
-package kr.binarybard.hireo.member.domain;
+package kr.binarybard.hireo.web.member.domain;
 
 public enum Role { FREELANCER, EMPLOYER, ADMINISTRATOR }
