+++ src/main/java/kr/binarybard/hireo/web/member/dto/MemberMapper.java
@@ -1,10 +1,10 @@
-package kr.binarybard.hireo.member.dto;
+package kr.binarybard.hireo.web.member.dto;
 
+import kr.binarybard.hireo.web.member.domain.Member;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
 
-import kr.binarybard.hireo.auth.dto.SignUpRequest;
-import kr.binarybard.hireo.member.domain.Member;
+import kr.binarybard.hireo.web.auth.dto.SignUpRequest;
 
 @Mapper(componentModel = "spring")
 public interface MemberMapper {
