+++ src/test/java/kr/binarybard/hireo/web/member/service/MemberServiceTest.java
@@ -1,4 +1,4 @@
-package kr.binarybard.hireo.member.service;
+package kr.binarybard.hireo.web.member.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -11,12 +11,11 @@ import static org.mockito.Mockito.when;
 import java.util.Optional;
 import kr.binarybard.hireo.web.auth.dto.SignUpRequest;
 import kr.binarybard.hireo.exception.MemberNotFoundException;
-import kr.binarybard.hireo.fixture.LoginFixture;
-import kr.binarybard.hireo.fixture.MemberFixture;
+import kr.binarybard.hireo.web.fixture.LoginFixture;
+import kr.binarybard.hireo.web.fixture.MemberFixture;
 import kr.binarybard.hireo.web.member.domain.Member;
 import kr.binarybard.hireo.web.member.dto.MemberMapper;
 import kr.binarybard.hireo.web.member.repository.MemberRepository;
-import kr.binarybard.hireo.web.member.service.MemberService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
